src: fix printf arg type mismatch (coverity)

+++ src/game/g_local.h
@@ -1522,7 +1522,7 @@ qboolean OnSameTeam(gentity_t *ent1, gentity_t *ent2);
 void reset_numobjectives(void);
 
 // g_mem.c
-void *G_Alloc(size_t size);
+void *G_Alloc(unsigned int size);
 void G_InitMemory(void);
 void Svcmd_GameMem_f(void);
